Update documentation, rename script, remove class
